news_cluster_service/requirements.txt
joblib
numpy
torch==1.11.0+cu113
sentence-transformers
sumy
kafka-python
pandas
pymongo